Fix __repr__() segfault and buggy items() on _Attributes
When calling attrs on a temporary instance, the tree may already be
deallocated when the tag name is being retrieved, which causes a null pointer
dereference. This fix doesn't really solve the problem, but at least it
prevents the crash.

Also fixes a bug in items() that made the method useless.

Fixes #39